pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "notion-portfolio-ops-template"
version = "0.1.0"
description = "Template for a Notion + OpenAI + GitHub Actions portfolio operations pipeline."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "openai>=1.75.0",
  "pydantic>=2.8.0",
  "python-dotenv>=1.0.1",
  "requests>=2.32.0",
]

[project.scripts]
portfolio-ops = "portfolio_sync.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

